+++ src/deploy.py
@@ -1,10 +1,14 @@
-from dotenv import dotenv_values
 from prefect import flow
+from prefect.runner.storage import GitRepository
+
+# Rappel : pour déployer, PREFECT_API_URL doit être non null, sinon on déploie
+# sur une instance éphémère.
 
 if __name__ == "__main__":
-    env = dotenv_values()
     flow.from_source(
-        source="https://github.com/ColinMaudry/decp-processing.git",
+        source=GitRepository(
+            url="https://github.com/ColinMaudry/decp-processing.git", branch="main"
+        ),
         entrypoint="src/flows.py:decp_processing",
     ).deploy(
         name="decp-processing",
@@ -12,6 +16,33 @@ if __name__ == "__main__":
         work_pool_name="local",
         ignore_warnings=True,
         cron="0 6 * * 1-5",
+        job_variables={
+            "env": {
+                "DECP_PROCESSING_PUBLISH": "True",
+                "DECP_DIST_DIR": "/srv/shared/decp/prod/dist",
+                "PREFECT_TASKS_REFRESH_CACHE": "False",
+            }
+        },
+    )
+
+    flow.from_source(
+        source=GitRepository(
+            url="https://github.com/ColinMaudry/decp-processing.git",
+            branch="dev",
+        ),
+        entrypoint="src/flows.py:decp_processing",
+    ).deploy(
+        name="decp-processing-dev",
+        description="Déploiement de la branche dev.",
+        work_pool_name="local",
+        ignore_warnings=True,
+        job_variables={
+            "env": {
+                "DECP_PROCESSING_PUBLISH": "False",
+                "DECP_DIST_DIR": "/srv/shared/decp/dev/dist",
+                "PREFECT_TASKS_REFRESH_CACHE": "True",
+            }
+        },
     )
 
     flow.from_source(
@@ -19,8 +50,72 @@ if __name__ == "__main__":
         entrypoint="src/flows.py:sirene_preprocess",
     ).deploy(
         name="sirene-preprocess",
-        description="Tous les mois, le 3",
+        description="Préparation des données SIRENE. Tous les mois, le 3",
         work_pool_name="local",
         ignore_warnings=True,
         cron="0 1 3 * *",
+        job_variables={
+            "env": {
+                "DECP_PROCESSING_PUBLISH": "True",
+                "DECP_DIST_DIR": "/srv/shared/decp/prod/dist",
+                "PREFECT_TASKS_REFRESH_CACHE": "False",
+            }
+        },
+    )
+
+    flow.from_source(
+        source=GitRepository(
+            url="https://github.com/ColinMaudry/decp-processing.git", branch="dev"
+        ),
+        entrypoint="src/flows.py:sirene_preprocess",
+    ).deploy(
+        name="sirene-preprocess-dev",
+        description="Préparation des données SIRENE.",
+        work_pool_name="local",
+        ignore_warnings=True,
+        job_variables={
+            "env": {
+                "DECP_PROCESSING_PUBLISH": "False",
+                "DECP_DIST_DIR": "/srv/shared/decp/dev/dist",
+                "PREFECT_TASKS_REFRESH_CACHE": "True",
+            }
+        },
+    )
+
+    flow.from_source(
+        source=GitRepository(
+            url="https://github.com/ColinMaudry/decp-processing.git", branch="main"
+        ),
+        entrypoint="src/flows.py:scrap_marches_securises",
+    ).deploy(
+        name="scrap-marches-securises",
+        description="Scraping des données de marches-securises.fr.",
+        ignore_warnings=True,
+        work_pool_name="local",
+        job_variables={
+            "env": {
+                "DECP_PROCESSING_PUBLISH": "True",
+                "DECP_DIST_DIR": "/srv/shared/decp/prod/dist",
+                "PREFECT_TASKS_REFRESH_CACHE": "False",
+            }
+        },
+    )
+
+    flow.from_source(
+        source=GitRepository(
+            url="https://github.com/ColinMaudry/decp-processing.git", branch="dev"
+        ),
+        entrypoint="src/flows.py:scrap_marches_securises",
+    ).deploy(
+        name="scrap-marches-securises-dev",
+        description="Scraping des données de marches-securises.fr.",
+        ignore_warnings=True,
+        work_pool_name="local",
+        job_variables={
+            "env": {
+                "DECP_PROCESSING_PUBLISH": "True",
+                "DECP_DIST_DIR": "/srv/shared/decp/dev/dist",
+                "PREFECT_TASKS_REFRESH_CACHE": "True",
+            }
+        },
     )