+++ src/flows.py
@@ -15,14 +15,16 @@ from config import (
     DATE_NOW,
     DECP_PROCESSING_PUBLISH,
     DIST_DIR,
+    MARCHES_SECURISES_SCRAPING_MODE,
     MAX_PREFECT_WORKERS,
+    MONTH_NOW,
     SIRENE_DATA_DIR,
     TRACKED_DATASETS,
 )
 from tasks.clean import clean_decp
 from tasks.dataset_utils import list_resources
 from tasks.enrich import enrich_from_sirene
-from tasks.get import get_resource
+from tasks.get import get_resource, scrap_marches_securises_month
 from tasks.output import generate_final_schema, save_to_databases, save_to_files
 from tasks.publish import publish_to_datagouv
 from tasks.transform import (
@@ -101,12 +103,13 @@ def decp_processing(enable_cache_removal: bool = False):
     ]
     dfs: list[pl.DataFrame] = [f.result() for f in futures if f.result() is not None]
 
-    create_table_artifact(
-        table=resources_artifact,
-        key="datagouvfr-json-resources",
-        description=f"Les ressources utilisées comme source ({DATE_NOW})",
-    )
-    del resources_artifact
+    if DECP_PROCESSING_PUBLISH:
+        create_table_artifact(
+            table=resources_artifact,
+            key="datagouvfr-json-resources",
+            description=f"Les ressources utilisées comme source ({DATE_NOW})",
+        )
+        del resources_artifact
 
     print("Fusion des dataframes...")
     df: pl.DataFrame = concat_decp_json(dfs)
@@ -126,8 +129,9 @@ def decp_processing(enable_cache_removal: bool = False):
         shutil.rmtree(DIST_DIR)
     os.makedirs(DIST_DIR)
 
-    print("Génération de l'artefact (statistiques) sur le base df...")
-    generate_stats(df)
+    if DECP_PROCESSING_PUBLISH:
+        print("Génération de l'artefact (statistiques) sur le base df...")
+        generate_stats(df)
 
     print("Génération du schéma et enregistrement des DECP aux formats CSV, Parquet...")
     df: pl.DataFrame = sort_columns(df, BASE_DF_COLUMNS)
@@ -139,7 +143,7 @@ def decp_processing(enable_cache_removal: bool = False):
     # Désactivé pour l'instant https://github.com/ColinMaudry/decp-processing/issues/124
     # make_data_tables()
 
-    if DECP_PROCESSING_PUBLISH.lower() == "true":
+    if DECP_PROCESSING_PUBLISH:
         print("Publication sur data.gouv.fr...")
         publish_to_datagouv()
     else:
@@ -174,5 +178,30 @@ def sirene_preprocess():
     print("☑️  Fin du flow sirene_preprocess.")
 
 
+@flow(log_prints=True)
+def scrap_marches_securises(mode=None, year=None):
+    mode = mode or MARCHES_SECURISES_SCRAPING_MODE
+
+    current_year = DATE_NOW[:4]
+
+    if mode == "month":
+        scrap_marches_securises_month(current_year, MONTH_NOW)
+
+    elif mode == "year":
+        year = year or current_year
+        for month in reversed(range(1, 13)):
+            month = str(month).zfill(2)
+            scrap_marches_securises_month(year, month)
+
+    elif mode == "all":
+        current_year = int(current_year)
+        for year in reversed(range(2018, current_year + 2)):
+            scrap_marches_securises("year", str(year))
+
+    else:
+        print("Mauvaise configuration")
+
+
 if __name__ == "__main__":
-    decp_processing()
+    # decp_processing()
+    scrap_marches_securises()