+++ src/tasks/clean.py
@@ -3,7 +3,7 @@ import re
 
 import polars as pl
 
-from config import DecpFormat
+from config import MARCHES_BAD_MONTANT, DecpFormat
 from tasks.transform import (
     explode_titulaires,
     process_modifications,
@@ -30,6 +30,16 @@ def clean_decp(lf: pl.LazyFrame, decp_format: DecpFormat) -> pl.LazyFrame:
     # Ajout du champ uid
     # TODO: à déplacer autre part, dans transform
     lf = lf.with_columns((pl.col("acheteur_id") + pl.col("id")).alias("uid"))
+
+    # Montants
+    # Certains marchés ont des montants qui posent problème, donc on les met à 1 euro
+    # ex : 221300015002472020F00075, 1.0E17
+    lf = lf.with_columns(
+        pl.when(pl.col("uid").is_in(MARCHES_BAD_MONTANT))
+        .then(pl.lit(1))
+        .otherwise(pl.col("montant"))
+        .alias("montant")
+    )
 
     # Dates
     date_replacements = {
@@ -72,6 +82,25 @@ def clean_decp(lf: pl.LazyFrame, decp_format: DecpFormat) -> pl.LazyFrame:
 
     # Explosion des titulaires
     lf = explode_titulaires(lf, decp_format)
+
+    # Remplacement des "" par null
+    lf = lf.with_columns(
+        pl.when(pl.col(pl.String).str.len_chars() == 0)
+        .then(None)
+        .otherwise(pl.col(pl.String))
+        .name.keep()
+    )
+
+    # Type identifiant = SIRET si vide (marches-securises.fr)
+    lf = lf.with_columns(
+        pl.when(
+            (pl.col("titulaire_typeIdentifiant").is_null())
+            & (pl.col("titulaire_id").str.len_chars() == 14)
+        )
+        .then(pl.lit("SIRET"))
+        .otherwise(pl.col("titulaire_typeIdentifiant"))
+        .alias("titulaire_typeIdentifiant")
+    )
 
     # NC
     lf = lf.with_columns(pl.col(pl.Utf8).replace("NC", None))