+++ src/tasks/get.py
@@ -1,18 +1,29 @@
+import json
 import tempfile
 from collections.abc import Iterator
 from functools import partial
 from pathlib import Path
+from time import sleep
 
+import httpx
 import ijson
 import orjson
 import polars as pl
+from bs4 import BeautifulSoup
 from httpx import stream
 from lxml import etree
 from prefect import task
 
-from config import DECP_FORMAT_2019, DECP_FORMATS, DIST_DIR, DecpFormat
+from config import (
+    DECP_FORMAT_2019,
+    DECP_FORMATS,
+    DECP_PROCESSING_PUBLISH,
+    DIST_DIR,
+    DecpFormat,
+)
 from tasks.clean import clean_invalid_characters, extract_innermost_struct
 from tasks.output import sink_to_files
+from tasks.publish import publish_scrap_to_datagouv
 from tasks.utils import gen_artifact_row, stream_replace_bytestring
 
 
@@ -62,8 +73,9 @@ def get_resource(
 
     # Ajout des stats de la ressource à l'artifact
     # https://github.com/ColinMaudry/decp-processing/issues/89
-    artifact_row = gen_artifact_row(r, lf, url, fields, decp_format)  # noqa
-    resources_artifact.append(artifact_row)
+    if DECP_PROCESSING_PUBLISH:
+        artifact_row = gen_artifact_row(r, lf, url, fields, decp_format)  # noqa
+        resources_artifact.append(artifact_row)
 
     # Exemple https://www.data.gouv.fr/datasets/5cd57bf68b4c4179299eb0e9/#/resources/bb90091c-f0cb-4a59-ad41-b0ab929aad93
     resource_web_url = (
@@ -110,10 +122,12 @@ def json_stream_to_parquet(
             use_float=True,
         )
 
-    tmp_file = tempfile.NamedTemporaryFile(mode="wb", suffix=".ndjson", delete=True)
+    tmp_file = tempfile.NamedTemporaryFile(mode="wb", suffix=".ndjson", delete=False)
 
     http_stream_iter = stream_get(url)
-    stream_replace_iter = stream_replace_bytestring(http_stream_iter, b"NaN,", b"null,")
+    stream_replace_iter = stream_replace_bytestring(
+        http_stream_iter, rb"NaN([,\n])", rb"null\1"
+    )
 
     # In first iteration, will find the right format
     chunk = next(stream_replace_iter)
@@ -237,3 +251,96 @@ def norm_titulaire(titulaire: dict):
     if "titulaire" in titulaire:
         titulaire = titulaire["titulaire"]
     return titulaire
+
+
+def get_html(url: str, client: httpx.Client) -> str or None:
+    headers = {
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:128.0) Gecko/20100101 Firefox/128.0",
+        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/png,image/svg+xml,*/*;q=0.8",
+        "Connection": "keep-alive",
+    }
+
+    def get_response() -> httpx.Response:
+        return client.get(url, timeout=timeout, headers=headers).raise_for_status()
+
+    timeout = httpx.Timeout(20.0, connect=60.0, pool=20.0, read=20.0)
+    try:
+        response = get_response()
+    except (httpx.ConnectError, httpx.ReadTimeout, httpx.HTTPStatusError):
+        print("3s break and retrying...")
+        sleep(3)
+        try:
+            response = get_response()
+        except (httpx.ConnectError, httpx.ReadTimeout, httpx.HTTPStatusError):
+            print("Skipped")
+            return None
+    html = response.text
+    return html
+
+
+# @task(
+#     cache_policy=INPUTS,
+#     persist_result=True,
+#     cache_expiration=datetime.timedelta(days=15),
+# )
+def get_json_marches_securises(url: str, client: httpx.Client) -> dict or None:
+    json_html_page = get_html(url, client)
+    sleep(0.1)
+    if json_html_page:
+        json_html_page = (
+            json_html_page.replace("</head>", "</head><body>") + "</body></html>"
+        )
+    else:
+        print("json_html_page is None, skipping...")
+        return None
+    json_html_page_soup = BeautifulSoup(json_html_page, "html.parser")
+    try:
+        decp_json = json.loads(json_html_page_soup.find("body").string)
+    except Exception as e:
+        print(json_html_page)
+        print(e)
+        return None
+    return decp_json
+
+
+@task(log_prints=True)
+def scrap_marches_securises_month(year: str, month: str):
+    marches = []
+    page = 1
+    with httpx.Client() as client:
+        while True:
+            search_url = (
+                f"https://www.marches-securises.fr/entreprise/?module=liste_donnees_essentielles&page={str(page)}&siret_pa=&siret_pa1=&date_deb={year}-{month}-01&date_fin={year}-{month}-31&date_deb_ms={year}-{month}-01&date_fin_ms={year}-{month}-31&ref_ume=&cpv_et=&type_procedure=&type_marche=&objet=&rs_oe=&dep_liste=&ctrl_key=aWwwS1pLUlFzejBOYitCWEZzZTEzZz09&text=&donnees_essentielles=1&search="
+                f"table_ms&"
+            )
+
+            def parse_result_page():
+                html_result_page = get_html(search_url, client)
+                soup = BeautifulSoup(html_result_page, "html.parser")
+                result_div = soup.find("div", attrs={"id": "liste_consultations"})
+                print("Year: ", year, "Month: ", month, "Page: ", str(page))
+                return result_div.find_all(
+                    "a", attrs={"title": "Télécharger au format Json"}
+                )
+
+            try:
+                json_links = parse_result_page()
+            except AttributeError:
+                sleep(3)
+                "Retrying result page download and parsing..."
+                json_links = parse_result_page()
+
+            if not json_links:
+                break
+            else:
+                page += 1
+            for json_link in json_links:
+                json_href = "https://www.marches-securises.fr" + json_link["href"]
+                decp_json = get_json_marches_securises(json_href, client)
+                marches.append(decp_json)
+        if len(marches) > 0:
+            dicts = {"marches": marches}
+            json_path = DIST_DIR / f"marches-securises_{year}-{month}.json"
+            with open(json_path, "w") as f:
+                f.write(json.dumps(dicts))
+            publish_scrap_to_datagouv(year, month, json_path)