+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+#### 2.1.1 2025-10-13
+
+- Stabilisation du scraping de marches-securises.fr
+- Amélioration du rendu des messages de release
+
+## 2.1.0 2025-10-13
+
+- scraping des données DECP de [marches-securises.fr](https://www.data.gouv.fr/datasets/donnees-essentielles-de-la-commande-publique-de-marches-securises-fr/) ([#111](https://github.com/ColinMaudry/decp-processing/issues/111))
+- ajout des données de marches-securises.fr aux données consolidées ([#111](https://github.com/ColinMaudry/decp-processing/issues/111))
+
+#### 2.0.5 - 2025-10-08
+
+- correction des NaN dans les données consolidées par le MINEF ([#127](https://github.com/ColinMaudry/decp-processing/issues/127))
+- auto-release à chaque fois que je push un tag
+
+#### 2.0.4 - 2025-10-04
+
+- nettoyage montant invalide de marché ([#125](https://github.com/ColinMaudry/decp-processing/issues/125))
+- publication du schéma au format [TableSchema](https://specs.frictionlessdata.io/table-schema) ([#126](https://github.com/ColinMaudry/decp-processing/issues/126))
+- amélioration des noms de colonnes dans le schéma pour les GUIs (`title`, `short_name`)
+
 #### 2.0.3 - 2025-09-29
 
 - correction de coquilles dans le schéma